Tidyup of getopt/usage + couple of codes/comments
No significant changes, but 'cleaned up' the code for the getopt long call, and matched parameter orders in the usage and the object constructor call (not sure why...). Tidied the usage text.

In a couple of places I did some minor re-formatting of the perl code to make it clearer. Added a couple of short comments.