Fixed problem with 'types' containing spaces. Some cleanup.
The old regexp didn't recognize types that have spaces in them. Here is a note from the GFF bible [1]:

        "Note that unescaped spaces are allowed within fields, meaning that parsers must split on tabs, not spaces."

[1] http://www.sequenceontology.org/resources/gff3.html

I noticed that SOFA doesn't have spaces in its terms, but it does have spaces in some of its term synonyms.

-  if ($source eq '' && $method =~ /^([\w-]+):([\w-]*)$/) {
-    $method = $1;
-    $source = $2;
+  if ($source eq '') {
+    ($method, $source) = split(/:/, $method, 2);

This was ballsing up the bp_seqfeature_delete.PLS script when searching for a type with a space.